requirements.txt
altgraph==0.17.3
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
easyocr==1.6.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
fonttools==4.38.0
greenlet==2.0.2
idna==3.4
imageio==2.25.0
itsdangerous==2.1.2
japanize-matplotlib==1.1.3
Jinja2==3.1.2
kiwisolver==1.4.4
macholib==1.16.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
ninja==1.11.1
numpy==1.24.1
opencv-python-headless==4.5.4.60
packaging==23.0
Pillow==9.4.0
pyclipper==1.3.0.post4
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.0
shapely==2.0.1
six==1.16.0
SQLAlchemy==2.0.0
tifffile==2023.1.23.1
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
